moved ramp command into separate package
