+++ src/main/java/frc/robot/commands/ramp/RampMove.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.cannon;
+package frc.robot.commands.ramp;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.subsystems.Ramp;
